third_party/blink/web_tests/vibration/vibration-detached-no-crash.html

<!DOCTYPE HTML>
<body>
<head>
<script src="../resources/js-test.js"></script>
<script src="vibration-utils.js"></script>
</head>
<script>
description("Accessing navigator.vibrate on a closed window.");

window.jsTestIsAsync = true;

var w;
function processMessage(event) {
    if (event.data == "opened") {
        w.close();
    } else if (event.data == "closed") {
        shouldBeFalse("w.navigator.vibrate([])");
        finishJSTest();
    }
}

if (window.testRunner) {
    testRunner.dumpAsText();
    testRunner.setPopupBlockingEnabled(false);
}
w = window.open('../resources/window-postmessage-open-close.html');
window.addEventListener("message", processMessage, false);
</script>
</body>
